High scores rework

+++ SimpleAndroidApp/src/org/ivan/simple/game/GameActivity.java
@@ -70,7 +70,7 @@ public class GameActivity extends Activity {
 		startActivity(intent);
     }
     
-    public void switchBackToChooseActivity(boolean complete, byte score) {
+    public void switchBackToChooseActivity(boolean complete, int score) {
 		Intent resultIntent = new Intent();
 		resultIntent.putExtra(LevelChooseActivity.LEVEL_COMPLETE, complete);
 		resultIntent.putExtra(LevelChooseActivity.COMPLETE_SCORE, score);
